Remove edge rate based dynamic isolation
There is currently not a good solution for dealing with seams between
two surfaces since we are considering all edge rates of the surface
together, this can lead to discrepancies.

This became more evident as we looked at content with sharpness.

+++ rtxmg/include/rtxmg/subdivision/subdivision_plan_hlsl.h
@@ -608,7 +608,7 @@ struct SubdivisionPlanContext
         return treeDescriptor;
     }
 
-    bool IsBSplinePatch(int level)
+    bool IsBSplinePatch(uint16_t level)
     {
         return GetTreeDescriptor().GetNumPatchPoints(level) == 0;
     }
